+++ src/main/java/com/dongjin/tastingnote/event/entity/UserEventType.java
@@ -1,0 +1,9 @@
+package com.dongjin.tastingnote.event.entity;
+
+public enum UserEventType {
+    NOTE_CREATED,
+    NOTE_RATED,
+    SEARCH,
+    VIEW_ALCOHOL,
+    VIEW_NOTE
+}